requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
diffusers==0.23.1
transformers==4.35.0
torch==2.1.0
accelerate==0.23.0
pydantic==2.4.2
pillow==10.1.0
xformers==0.0.21
opencv-python==4.8.1.78
numpy==1.24.3
insightface==0.7.3
onnxruntime-gpu==1.15.1
timm==0.9.7
controlnet-aux==0.0.6
open-clip-torch==2.20.0
torchvision==0.16.0
scipy==1.11.3
safetensors==0.4.0
huggingface_hub==0.19.4
tqdm==4.66.1
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
realesrgan==0.3.0 